@@ -31,7 +31,7 @@ class TracksController < ApplicationController
   end
 
   def update
-    if @track.update_attributes(track_params)
+    if @track.update(track_params)
       redirect_to conference_program_tracks_path(conference_id: @conference.short_title),
                   notice: 'Track request successfully updated.'
     else
