@@ -17,7 +17,7 @@ describe Admin::VersionsController do
       end
 
       it 'reverts all changes for update actions' do
-        conference.update_attributes(short_title: 'testtitle', description: 'Some random text')
+        conference.update(short_title: 'testtitle', description: 'Some random text')
         get :revert_object, params: { id: conference.versions.last.id }
         conference.reload
         expect(conference.short_title).to eq 'exampletitle'
@@ -62,7 +62,7 @@ describe Admin::VersionsController do
       end
 
       it 'reverts specified change for update actions' do
-        conference.update_attributes(short_title: 'testtitle', description: 'Some random text')
+        conference.update(short_title: 'testtitle', description: 'Some random text')
         get :revert_attribute, params: { id: conference.versions.last.id, attribute: 'short_title' }
         conference.reload
         expect(conference.short_title).to eq 'exampletitle'
@@ -70,8 +70,8 @@ describe Admin::VersionsController do
       end
 
       it 'shows correct flash on trying to revert to the current state' do
-        conference.update_attributes(short_title: 'testtitle', description: 'Some random text')
-        conference.update_attributes(short_title: 'exampletitle')
+        conference.update(short_title: 'testtitle', description: 'Some random text')
+        conference.update_attribute(:short_title, 'exampletitle')
         get :revert_attribute, params: { id: conference.versions[-2].id, attribute: 'short_title' }
         expect(flash[:error]).to match('The item is already in the state that you are trying to revert it back to')
         expect(conference.short_title).to eq 'exampletitle'
@@ -79,7 +79,7 @@ describe Admin::VersionsController do
 
       it 'fails on trying to revert deleted object' do
         event_type = conference.program.event_types.first
-        event_type.update_attributes(title: 'New Event Title')
+        event_type.update_attribute(:title, 'New Event Title')
         event_type.destroy
         get :revert_attribute, params: { id: event_type.versions[-2].id, attribute: 'title' }
         conference.reload
@@ -93,7 +93,7 @@ describe Admin::VersionsController do
       end
 
       it 'revert fails when attribute is invalid' do
-        conference.update_attributes(short_title: 'testtitle', description: 'Some random text')
+        conference.update(short_title: 'testtitle', description: 'Some random text')
         before_conference_title = conference.title
         # Note: even though title is a valid attribute of conference, it was not updated in the change we are trying to revert
         get :revert_attribute, params: { id: conference.versions.last.id, attribute: 'title' }
@@ -117,12 +117,12 @@ describe Admin::VersionsController do
         before :each do
           @user = create(:user)
 
-          conference.update_attributes(short_title: 'testtitle', description: 'Some random text')
+          conference.update(short_title: 'testtitle', description: 'Some random text')
           @version_organizer = conference.versions.last
           cfp = create(:cfp, program: conference.program)
           @version_cfp = cfp.versions.last
           registration = create(:registration, conference: conference)
-          registration.update_attributes(attended: true)
+          registration.update_attribute(:attended, true)
           @version_info_desk = registration.versions.last
         end
 
