@@ -68,10 +68,6 @@ describe EmailSettings do
     }
   end
 
-  it 'has a valid factory' do
-    expect(build(:email_settings)).to be_valid
-  end
-
   describe '#get_values' do
     context 'user has name' do
       it 'returns correct key-value pairs' do
@@ -81,7 +77,7 @@ describe EmailSettings do
 
     context 'user does not have name' do
       before do
-        user.update_attributes(name: nil)
+        user.update(name: nil)
         username_hash = { 'name' => 'johnd' }
         expected_hash.merge!(username_hash)
       end
@@ -108,7 +104,7 @@ describe EmailSettings do
 
     context 'conference has venue' do
       before do
-        conference.update_attributes(venue: create(:venue))
+        conference.venue = create(:venue)
         venue_hash = { 'venue' => conference.venue.name, 'venue_address' => conference.venue.address }
         expected_hash.merge!(venue_hash)
       end
@@ -120,9 +116,9 @@ describe EmailSettings do
 
     context 'conference has registration period' do
       before do
-        conference.update_attributes(registration_period: create(:registration_period,
-                                                                 start_date: Date.new(2014, 05, 03),
-                                                                 end_date:   Date.new(2014, 05, 05)))
+        conference.registration_period = create(:registration_period,
+                                                start_date: Date.new(2014, 05, 03),
+                                                end_date:   Date.new(2014, 05, 05))
         registration_period_hash = { 'registration_start_date' => Date.new(2014, 05, 03), 'registration_end_date' => Date.new(2014, 05, 05) }
         expected_hash.merge!(registration_period_hash)
       end
