@@ -180,9 +180,15 @@ module FormatHelper
       strikethrough:       true,
       footnotes:           true,
       superscript:         true
+      # autolink:                     true,
+      # space_after_headers:          true,
+      # no_intra_emphasis:            true,
+      # fenced_code_blocks:           true,
+      # disable_indented_code_blocks: true,
+      # safe_links_only:              true
     }
     markdown = Redcarpet::Markdown.new(Redcarpet::Render::HTML.new(escape_html: escape_html), options)
-    markdown.render(text).html_safe
+    sanitize(markdown.render(text))
   end
 
   def markdown_hint(text='')
