@@ -15,7 +15,7 @@ describe Api::V1::RoomsController do
         get :index, params: { format: :json }
         json = JSON.parse(response.body)['rooms']
 
-        expect(response).to be_success
+        expect(response).to be_successful
 
         expect(json.pluck('name')).to contain_exactly('Conference Room', 'Test Room')
       end
@@ -27,7 +27,7 @@ describe Api::V1::RoomsController do
         get :index, params: { conference_id: conference.short_title, format: :json }
         json = JSON.parse(response.body)['rooms']
 
-        expect(response).to be_success
+        expect(response).to be_successful
 
         expect(json.length).to eq(1)
         expect(json[0]['name']).to eq('Conference Room')
