@@ -14,7 +14,7 @@ describe Api::V1::TracksController do
         get :index, params: { format: :json }
         json = JSON.parse(response.body)['tracks']
 
-        expect(response).to be_success
+        expect(response).to be_successful
 
         expect(json.pluck('name')).to contain_exactly('Conference Track', 'Test Track')
       end
@@ -26,7 +26,7 @@ describe Api::V1::TracksController do
         get :index, params: { conference_id: conference.short_title, format: :json }
         json = JSON.parse(response.body)['tracks']
 
-        expect(response).to be_success
+        expect(response).to be_successful
 
         expect(json.length).to eq(1)
         expect(json[0]['name']).to eq('Conference Track')
