@@ -8,129 +8,58 @@ feature Commercial do
   let!(:cfp) { create(:cfp, program: conference.program) }
   let!(:organizer) { create(:organizer, resource: conference) }
   let!(:participant) { create(:user) }
-
-  context 'in admin area' do
-
-    before do
-      sign_in organizer
-      visit admin_conference_commercials_path(conference.short_title)
-    end
-
-    scenario 'contains the conference name in the title', feature: true, js: true do
-      header = conference.title + ' Materials'
-      expect(page).to have_content(header)
-    end
-
-    scenario 'adds, updates, deletes of a conference', feature: true, js: true do
-      # Create valid commercial
-      fill_in 'commercial_title', with: 'My Cool Talk!'
-      fill_in 'commercial_url', with: 'https://www.youtube.com/watch?v=M9bq_alk-sw'
-      click_button 'Save Materials'
-      page.find('#flash')
-      expect(flash).to eq('Materials were successfully created.')
-      page.find('#flash .button.close').click
-      expect(conference.commercials.count).to eq(1)
-
-      commercial = conference.commercials.where(url: 'https://www.youtube.com/watch?v=M9bq_alk-sw').first
-
-      fill_in "commercial_url_#{commercial.id}", with: 'https://www.youtube.com/watch?v=VNkDJk5_9eU'
-      click_button 'Update'
-      page.find('#flash')
-      expect(flash).to eq('Materials were successfully updated.')
-      page.find('#flash .button.close').click
-      expect(conference.commercials.count).to eq(1)
-      commercial.reload
-      expect(commercial.url).to eq 'https://www.youtube.com/watch?v=VNkDJk5_9eU'
-
-      # Delete commercial
-      page.accept_alert do
-        click_link 'Delete'
-      end
-      page.find('#flash')
-      expect(flash).to eq('Materials were successfully removed.')
-      expect(conference.commercials.count).to eq(0)
-    end
+  let!(:event) { create(:event, program: conference.program, title: 'Example Proposal') }
+  let!(:event_user) do
+    create(:event_user, user: participant, event: event, event_role: 'submitter')
   end
 
-  context 'in public area' do
-    let!(:event) { create(:event, program: conference.program, title: 'Example Proposal') }
-    let!(:event_user) do
-      create(:event_user, user: participant, event: event, event_role: 'submitter')
-    end
+  before(:each) do
+    sign_in participant
+  end
 
-    before(:each) do
-      sign_in participant
-    end
+  scenario 'adds a valid commercial of an event', feature: true, js: true do
+    visit edit_conference_program_proposal_path(conference.short_title, event.id)
+    click_link 'Materials'
+    fill_in 'commercial_url', with: 'https://www.youtube.com/watch?v=M9bq_alk-sw'
 
-    scenario 'adds a valid commercial of an event', feature: true, js: true do
-      visit edit_conference_program_proposal_path(conference.short_title, event.id)
-      click_link 'Materials'
+    # Workaround to enable the 'Create Commercial' button
+    page.execute_script("$('#commercial_submit_action').prop('disabled', false)")
+
+    click_button 'Create Materials'
+    page.find('#flash')
+    expect(flash).to eq('Materials were successfully created.')
+  end
+
+  scenario 'updates a commercial of an event', feature: true, js: true do
+    commercial = create(:commercial,
+                        commercialable_id:   event.id,
+                        commercialable_type: 'Event')
+    visit edit_conference_program_proposal_path(conference.short_title, event.id)
+    click_link 'Materials'
+    within('.thumbnail') do
       fill_in 'commercial_url', with: 'https://www.youtube.com/watch?v=M9bq_alk-sw'
-
-      # Workaround to enable the 'Save Materials' button
+      # Workaround to enable the 'Create Commercial' button
       page.execute_script("$('#commercial_submit_action').prop('disabled', false)")
-
-      click_button 'Save Materials'
-      page.find('#flash')
-      expect(flash).to eq('Materials were successfully created.')
+      click_button 'Update Materials'
     end
+    page.find('#flash')
+    expect(flash).to eq('Materials were successfully updated.')
+    expect(event.commercials.count).to eq(1)
+    commercial.reload
+    expect(commercial.url).to eq('https://www.youtube.com/watch?v=M9bq_alk-sw')
+  end
 
-    scenario 'does not add an invalid commercial of an event', feature: true, js: true do
-      # TODO-SNAPCON
-      skip('Snap!Con allows all materials to be saved.')
-      visit edit_conference_program_proposal_path(conference.short_title, event.id)
-      click_link 'Materials'
-      fill_in 'commercial_url', with: 'invalid_commercial_url'
-      expect(page).to have_content('No embeddable content')
-      expect(page).to have_css("button[type='submit']:disabled", text: 'Save Materials')
+  scenario 'deletes a commercial of an event', feature: true, js: true do
+    create(:commercial,
+           commercialable_id:   event.id,
+           commercialable_type: 'Event')
+    visit edit_conference_program_proposal_path(conference.short_title, event.id)
+    click_link 'Materials'
+    page.accept_alert do
+      click_link 'Delete'
     end
-
-    scenario 'updates materials of an event', feature: true, js: true do
-      commercial = create(:commercial,
-                          commercialable_id:   event.id,
-                          commercialable_type: 'Event')
-      visit edit_conference_program_proposal_path(conference.short_title, event.id)
-      click_link 'Materials'
-      fill_in "commercial_title_#{commercial.id}", with: 'My Cool Talk!'
-      fill_in "commercial_url_#{commercial.id}", with: 'https://www.youtube.com/watch?v=M9bq_alk-sw'
-      click_button 'Update'
-      page.find('#flash')
-      expect(flash).to eq('Materials were successfully updated.')
-      expect(event.commercials.count).to eq(1)
-      commercial.reload
-      expect(commercial.title).to eq('My Cool Talk!')
-      expect(commercial.url).to eq('https://www.youtube.com/watch?v=M9bq_alk-sw')
-    end
-
-    scenario 'does not update a commercial of an event with invalid data', feature: true do
-      commercial = create(:commercial,
-                          commercialable_id:   event.id,
-                          commercialable_type: 'Event',
-                          url:                 'https://www.youtube.com/watch?v=BTTygyxuGj8')
-      visit edit_conference_program_proposal_path(conference.short_title, event.id)
-      click_link 'Materials'
-      fill_in 'commercial_title', with: 'My Cool Talk!'
-      fill_in "commercial_url_#{commercial.id}", with: 'invalid_commercial_url'
-      click_button 'Update'
-      find('#flash')
-      expect(current_path).to eq edit_conference_program_proposal_path(conference.short_title, event.id)
-      expect(flash).to include('An error prohibited materials from being saved:')
-      commercial.reload
-      expect(commercial.url).to eq('https://www.youtube.com/watch?v=BTTygyxuGj8')
-    end
-
-    scenario 'deletes a commercial of an event', feature: true, js: true do
-      create(:commercial,
-             commercialable_id:   event.id,
-             commercialable_type: 'Event')
-      visit edit_conference_program_proposal_path(conference.short_title, event.id)
-      click_link 'Materials'
-      page.accept_alert do
-        click_link 'Delete'
-      end
-      page.find('#flash')
-      expect(flash).to eq('Materials were successfully destroyed.')
-      expect(event.commercials.count).to eq(0)
-    end
+    page.find('#flash')
+    expect(flash).to eq('Materials were successfully destroyed.')
+    expect(event.commercials.count).to eq(0)
   end
 end
