@@ -27,14 +27,8 @@ describe Payment do
   end
 
   describe 'validations' do
-    it 'has a valid factory' do
-      expect(build(:payment)).to be_valid
-    end
-
     it { is_expected.to validate_presence_of(:status) }
-
     it { is_expected.to validate_presence_of(:user_id) }
-
     it { is_expected.to validate_presence_of(:conference_id) }
   end
 
@@ -133,7 +127,7 @@ describe Payment do
 
       context 'when the request to Stripe is invalid' do
         it 'raises exception' do
-          StripeMock.prepare_error(Stripe::InvalidRequestError.new('Your request is invalid.', code: 402))
+          StripeMock.prepare_error(Stripe::InvalidRequestError.new('Your request is invalid.', {}, code: 402))
           expect{ payment.purchase }.not_to raise_error
         end
       end
