@@ -20,10 +20,6 @@ require 'spec_helper'
 describe TicketPurchase do
 
   describe 'validations' do
-    it 'has a valid factory' do
-      expect(build(:ticket_purchase)).to be_valid
-    end
-
     it 'is not valid without a conference_id' do
       should validate_presence_of(:conference_id)
     end
@@ -56,9 +52,7 @@ describe TicketPurchase do
       let(:registration_ticket) { create(:registration_ticket) }
       let(:ticket_purchase) { build(:ticket_purchase, ticket: registration_ticket, quantity: 1) }
 
-      it 'it is valid, if quantity for registration tickets is less than or equal to one' do
-        expect(ticket_purchase.valid?).to eq true
-      end
+      it { is_expected.to validate_numericality_of(:quantity) }
 
       it 'it is not valid, if quantity for registration tickets is greater than to one' do
         ticket_purchase.quantity = 4
